Move db.py to core/db/reminders_db.py
